requirements.txt
opencv-python~=4.11.0.86
numpy~=2.2.2
pytest~=8.3.4
PyYAML~=6.0.2